refactor: split preprocessor into modules
